Move ambassador data
