+++ src/pages/community/ambassadors/index.mdx
@@ -4,7 +4,7 @@ description: An introduction to the GraphQL Ambassador Program
 
 import { Button } from "../../../app/conf/_components/button"
 import { AmbassadorGrid } from "../../../components/ambassador-grid"
-import {ambassadors} from "./data"
+import {ambassadors} from "../../../components/info-card/ambassador-data"
 
 # GraphQL Ambassador Program
